Address some review feedback
